fix(dtypes): let the lumped-RLC, NTFF and 2-D TFSF scan carries follow the ambient precision instead of a hard pin (closes #646)
With `jax_enable_x64` on, three `lax.scan` carry families raised
`scan body function carry input and carry output must have equal types`,
reddening 9 fast-suite tests. Reproduced on a pinned `git archive` of the
parent commit: `JAX_ENABLE_X64=0` -> 21 passed, `=1` -> 9 failed / 12 passed,
matching the issue exactly.

Corrected mechanism. The issue attributed the promotion to "the field
arithmetic they are driven from", i.e. the field dtype. Measured, that is not
it: under x64 the fields stay float32 and no `fdtd` carry component appears in
any mismatch list. The promoting quantity is `grid.dt`, which `Grid.__init__`
computes with numpy and so is an **np.float64 scalar**. numpy scalars are
STRONGLY typed in JAX, so `complex64_array * np.float64` is complex128 under
x64 while `complex64_array * <python float>` stays complex64. With x64 off JAX
clamps float64 to float32 and the pins happened to agree. This matters for the
fix: `promote_types(field_dtype, float32)` at the ALLOCATION alone does not
close these scans, because with float32 fields it yields complex64 while the
body still returns complex128. The allocation policy has to be paired with
pinning the body's arithmetic to the dtype the carry was allocated at.

A third site, absent from the issue's two-row table: the 2-D oblique TFSF aux
grid (`rfx/sources/tfsf_2d.py`), on
`carry['tfsf'].{ez_2d,hx_2d,hy_2d,psi_ez_x*,psi_hy_x*}`. It accounts for the
9th failure (`test_compiled_runner_tfsf_oblique_matches_manual_loop`).
Measured split of the nine: 3 lumped-RLC, 5 NTFF, 1 TFSF-2D.

Dtype policy, unchanged from #630/#644: `promote_types(field_dtype, float32)`
-- promote, never pin, never a bare copy of the field dtype. Both failure
modes are live and are now pinned by tests: a flat float32 pin breaks the #404
oblique-Bloch path, which needs a genuinely complex carry; naively following
the field dtype puts a recursive accumulator in float16 under
`precision="mixed"`.

Per site:

* `rfx/farfield.py` -- new `ntff_accum_dtype()` is the single policy shared by
  `init_ntff_data` (allocation, now taking `field_dtype=`) and
  `accumulate_ntff` (which pins its phase arithmetic to the allocated dtype).
  float16->complex64, float32->complex64, float64->complex128,
  complex64->complex64. `rfx/simulation.py` threads `_field_dtype` in, beside
  the CPML `psi` carry. The `nonuniform`/`subgridded` callers need no change:
  the body now derives its dtype from the carry, so their scans close too.
* `rfx/lumped.py` -- `init_rlc_state()`'s hard `dtype=jnp.float32` default
  becomes the ambient default float with a float32 floor. Note the production
  `run()` path was ALREADY correct (`rlc_carry_dtype` threads the metas'
  np.float64 dtypes); only the bare-default callers were broken. Measured, not
  assumed: `test_lumped_rlc_scan_closes_under_x64` passes on the parent commit
  while `test_rlc_carry_follows_ambient_float_with_a_float32_floor` fails.
* `rfx/sources/tfsf_2d.py` -- the aux state is narrowed back to the carry's
  dtype on OUTPUT rather than by casting dx/dt/t on input. Deliberate: those
  scalar precomputes (`dt/MU_0`, `exp(-1j*k_y*dx)`, the source term) are
  evaluated in float64/complex128 today and rounded once where they meet the
  complex64 arrays. Casting inputs first moves the rounding earlier
  (`float32(dt)/MU_0` instead of `float32(dt/MU_0)`) and perturbs the default
  path by an ULP; narrowing the output is a no-op with x64 off and so is
  bit-identical by construction.

Also fixed on the same path, same class as #644's second defect:
`apply_tfsf_2d_e/h` scatter-added a float64/complex128 product into a
complex64 field, tripping JAX's "cannot safely cast" FutureWarning (a future
hard error). The correction is now narrowed to the target field's dtype.

Evidence.

- x64 matrix over the issue's six node sets, pinned `git archive` of parent vs
  this branch: parent 21 passed / 9 failed (x64 0/1); branch 21 passed at BOTH
  env values. With the new file included: 35 passed at both.
- Production bit-identity as an identity claim, no tolerances: SHA-256 of raw
  field bytes at the default `precision="float32"` matches the parent commit
  on 7/7 fixtures -- pec, cpml, lossy+cpml, nu+cpml, ntff/rcs, lumped-rlc, and
  an oblique-TFSF fixture added because the tfsf_2d change would otherwise be
  unhashed.
- Harness sensitivity AND selectivity demonstrated, so the 7/7 is not a
  harness that cannot fail: a +3 ppm `dx` nudge moves all 7; a +3 ppm
  perturbation confined to the NTFF phase moves ONLY ntff_rcs; to the RLC
  ADE coefficient, ONLY lumped_rlc; to the 2-D aux source time, ONLY
  oblique_tfsf. Each changed lane is covered by exactly one fixture.
- Suites: lumped+mixed-precision 61 passed, rcs/ntff/farfield 50 passed,
  bloch/oblique 34 passed, boundary 87 passed, scoped-x64 neighbours 19
  passed. ruff clean over rfx/, tests/, conftest.py.

Pre-existing failure, NOT caused here and not fixed here:
`test_oblique_rcs_absolute_sigma.py::test_po_oracle_matches_closed_form` dies
on `np.trapz`, removed in numpy 2.x. Identical failure on the pinned parent
commit; environment drift, worth its own issue.

conftest.py grows an autouse guard that fails the test which leaves
`jax_enable_x64` flipped, so a future leak is blamed on its cause rather than
on the next test in the shard (the PR #645 failure mode). Two deviations from
the issue's sketch, both deliberate: it is function-scoped, because a
session-scoped teardown fires once at the end and cannot name the culprit; and
it compares against the session's initial value rather than a hard False,
because the supported way to reproduce this issue is to run the suite under
`JAX_ENABLE_X64=1`, where asserting False would fail everything. It also
restores the flag before asserting, so a leak costs exactly one red test --
verified with a deliberate leaker: with the restore its neighbour still
passes, without it both error.

Coverage boundary, stated rather than glossed: the NU lane and the NTFF lane
are hashed separately but not in combination; that pairing is covered
functionally by tests/test_farfield_nonuniform.py, not by a byte comparison.

Corrects two docstrings that were measurably wrong and would mislead the next
reader: `rlc_carry_dtype` claimed the metas are plain Python floats and that it
returns float32 (they are np.float64 and it returns float64; the float32 came
from JAX clamping at allocation), and `init_ntff_data` justified its complex64
pin with "x64 caused an MLIR scan-carry mismatch, issue #14" -- causality
backwards, the pin is what raises the mismatch.

R3: memory=docs/agent-memory/rfx-known-issues.md "Added 2026-08-11 -- #630
RESOLVED" ("jnp.promote_types(state.ex.dtype, jnp.float32) collapsed EVERY
per-stage anomaly"; "at float32 field storage the fix is a no-op there") --
consistent, this extends the same promote-don't-pin policy to three more
carries and re-measures the no-op | R2-attempts=1 | falsifier=re-run the
7-fixture SHA-256 harness at x64=0, parent vs branch; any differing hash
falsifies the bit-identity claim, and RFX646_PERTURB=ntff must move ntff_rcs
alone

+++ rfx/simulation.py
@@ -852,7 +852,11 @@ def _build_step_setup(
         from rfx.farfield import accumulate_ntff as _accumulate_ntff
         init_ntff_data_fn = _init_ntff_data
         accumulate_ntff_fn = _accumulate_ntff
-        carry_init["ntff"] = _init_ntff_data(ntff)
+        # NTFF accumulator dtype follows the field dtype with a float32 floor
+        # (#646), mirroring the CPML psi carry above. accumulate_ntff pins its
+        # phase arithmetic to whatever is allocated here, so the scan carry
+        # closes under jax_enable_x64 instead of raising a dtype mismatch.
+        carry_init["ntff"] = _init_ntff_data(ntff, field_dtype=_field_dtype)
 
     if use_dft_planes:
         carry_init["dft_planes"] = tuple(probe.accumulator for probe in dft_planes)
